# backend/data-service/models/set.go
package models

type Set struct {
	SetNum         string  `gorm:"primaryKey" json:"set_num"`
	Name           string  `gorm:"size:255;not null" json:"name"`
	Year           int     `gorm:"not null" json:"year"`
	ThemeID        int     `json:"theme_id"` // RebrickableID из Series
	Series         *Series `gorm:"foreignKey:ThemeID;references:RebrickableID" json:"series,omitempty"`
	NumParts       int     `json:"num_parts"`
	SetImgURL      string  `json:"set_img_url"`
	SetURL         string  `json:"set_url"`
	LastModifiedDT string  `json:"last_modified_dt"`
}